requirements.txt
requests
ccxt == 1.45.18
tweepy
python-dotenv
